added endpoints of added specializations and services to doctor

+++ backend/src/main/java/com/personal/patient/account/repositories/DoctorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface DoctorRepository extends CrudRepository<Doctor, Integer> {
+public interface DoctorRepository extends CrudRepository<Doctor, Long> {
     List<Doctor> findByHospital(Hospital hospital);
 }
